move LibGit2 to stdlib
