+++ doc/src/index.md
@@ -115,7 +115,6 @@ Please read the [release notes](NEWS.md) to see what has changed since the last 
       * [Bounds checking](@ref)
       * [Proper maintenance and care of multi-threading locks](@ref)
       * [Arrays with custom indices](@ref)
-      * [Base.LibGit2](@ref)
       * [Module loading](@ref)
       * [Inference](@ref)
   * Developing/debugging Julia's C code
