+++ stdlib/LibGit2/src/callbacks.jl
