+++ stdlib/LibGit2/src/gitcredential.jl
