+++ stdlib/LibGit2/src/merge.jl
