+++ stdlib/LibGit2/src/oid.jl
