+++ stdlib/LibGit2/src/reference.jl
