+++ stdlib/LibGit2/src/strarray.jl
