+++ stdlib/LibGit2/test/keys/valid-passphrase.pub
