+++ stdlib/LibGit2/test/runtests.jl
@@ -1,0 +1,2 @@
+include("libgit2.jl")
+include("online.jl")
