+++ stdlib/Pkg/src/cache.jl
@@ -2,8 +2,8 @@
 
 module Cache
 
-import Pkg
-import ...LibGit2, ..Dir, ..PkgError
+import LibGit2
+import ..Dir, ..PkgError
 using ..Types
 
 rewrite_url_to = "https"
